Deleted ShaderParametersSSAO
